==> 2_data_cleaning_covid.ipynb <==
### 90803 Data Cleaning and Question Definition
### Data Cleaning: COVID-19 Datasets

**Team 14**

Chi-Shiun Tsai & Colton Lapp

This notebook is used for cleaning the COVID-19 datasets from the NYTimes.

### 0. Importing libraries

import numpy as np
import pandas as pd
from datetime import datetime
from census import Census
from us import states

### 1. Reading datasetss
We tried to configure this to work with google drive but it wasn't working well. You may have to download the file manually if you want to recreate this notebook from scratch at the link below

# COVID-19 data

#Try to read in from google drive link:
try:
    covid_20 = pd.read_csv('https://drive.google.com/uc?export=download&id=1anYHHwpp1gISwfRaIBu3xgZmoIeGd9aD',  sep=',',lineterminator='\n', dtype={'fips': str})
except:
    print("Could not read covid data from google drive link. Please manually download from this link:\n \
            https://drive.google.com/uc?export=download&id=1anYHHwpp1gISwfRaIBu3xgZmoIeGd9aD \n \
            and read in below manually")
    
if 'covid_20' not in locals():
    try:
        covid_20 = pd.read_csv('../data/COVID-19 Cases/us-counties-2020.csv', dtype={'fips': str})
    except:
        print("Please download file in correct directory")

#### Merge Population Data in with Covid data on the county level using Census API

# Population data

# Initialize Census object using census API key
census_object = Census("bf6690c63fb4bbd43ccae839241e9be45bfc0881")

# 2020 population data
pop_2020 = census_object.acs5.state_county(fields = ('NAME', 'B01003_001E'),
                                      state_fips = "*",
                                      county_fips = "*",
                                      year = 2020)

#Print
pop_2020_df = pd.DataFrame(pop_2020)
pop_2020_df

#### Create county fips code by combining state and county fips codes

pop_2020_df['fips'] = pop_2020_df['state'] + pop_2020_df['county']
pop_2020_df

We want to focus on the aggregated year level. Therefore, we want to keep only the number of cumulative cases and deaths for the last day of 2020.

def keep_last(df):
    df['date'] = pd.to_datetime(df['date'])
    df.sort_values('date', inplace=True, ascending=False)
    df = df.groupby(by=['county', 'state'], as_index=False).first()
    df['year'] = df['date'].dt.year
    df.drop(columns=['date'], inplace=True)
    return df

covid = keep_last(covid_20)
covid['fips'] = covid['fips'].astype(str)
covid

### 2. Data cleaning

# Check for missing values
covid.isnull().sum(axis=0)

covid[covid['deaths'].isnull()]['state'].unique()

In this dataset, Puerto Rico does not record number of deaths. We can drop Puerto Rico since we will focus on the contiguous United States.

covid.drop(covid[covid['deaths'].isnull()].index, inplace=True)
covid

covid[covid['fips'] == 'None']['county'].unique()

# Join two dataframes
covid_df = covid.merge(pop_2020_df, how='left', on='fips')
covid_df

#### Investigate counties without fips codes

covid_df[covid_df['fips'] == 'None']['county_x'].unique()

# Drop unknown counties
covid_df.drop(covid_df[covid_df['county_x'] == 'Unknown'].index, inplace=True)

#### Fill in county code for MSA's that have code missing
We will use a county that is within the MSA

# Joplin population 2020: 51703
# Kansas city: 507932
# New York city: 8804190

# Fill in missing population values
covid_df.loc[covid_df['county_x'] == 'Joplin', 'B01003_001E'] = 51703
covid_df.loc[covid_df['county_x'] == 'Kansas City', 'B01003_001E'] = 507932
covid_df.loc[covid_df['county_x'] == 'New York City', 'B01003_001E'] = 8804190

# Check
covid_df[covid_df['county_x'] == 'Joplin']

# Calculate cases and deaths per capita

covid_df['cases'] = covid_df['cases']/covid_df['B01003_001E']
covid_df['deaths'] = covid_df['deaths']/covid_df['B01003_001E']

covid_df

covid_df[covid_df['fips'] == 'None']

### Fill in missing fips values

# Jasper County: 13159
# Jackson county: 29095
# New York County: 36061

covid_df.loc[covid_df['county_x'] == 'Joplin', 'fips'] = 13159
covid_df.loc[covid_df['county_x'] == 'Kansas City', 'fips'] = 29095
covid_df.loc[covid_df['county_x'] == 'New York City', 'fips'] = 36061

# Drop unnecessary columns for final dataframe
covid_df_cleaned = covid_df[['county_x', 'state_x', 'fips', 'cases', 'deaths', 'year']]
covid_df_cleaned.columns = ['county', 'state', 'fips', 'cases', 'deaths', 'year']
covid_df_cleaned

### 3. Saving cleaned dataset

covid_df_cleaned.to_csv('../data/data_cleaned/covid.csv', index=False)

### References

* Data source: https://github.com/nytimes/covid-19-data
* https://pandas.pydata.org/docs/reference/api/pandas.isnull.html
