==> 4_data_cleaning_crime.ipynb <==
### 90803 Data Cleaning and Question Definition
# Data Cleaning: Crime Data

**Team 14**

Chi-Shiun Tsai & Colton Lapp

This notebook is used for cleaning the crime data from FBI.

### 0. Importing libraries

import glob
import numpy as np
import pandas as pd
from datetime import datetime
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
import os

### 1. Reading datasets

Read all excel files and merge them into a single dataframe.


try:
    # Get the path of all Excel files 
    files = glob.glob("../data/State Tables Offenses by Agency 2020/*.xls")
    
    apd_dta = []
    for f in files:
        df = pd.read_excel(f, skiprows=3)
        col_names = df.iloc[0] 
        df = df[1:] 
        df.columns = col_names
        df.columns = df.columns.str.replace('\n', '')

        # Fill in missing values and subset only agency of cities
        df['Agency Type'] = df['Agency Type'].ffill(axis = 0)
        df = df[df['Agency Type']=='Cities']
        df.drop(['Agency Type'], axis=1, inplace=True)

        # Keep only first columns
        df = df.iloc[:, :3]
        df.sort_index(inplace=True) 

        # Add column for state with filename
        s = "/".join(f.split("/")[-1:]) 
        s = s.rstrip("/").split('_Offense')[0]
        df['State'] = s
        apd_dta.append(df)

    df_merged = pd.concat(apd_dta, ignore_index=True)

except:
    print("Error reading in raw crime data. \
These files are small so they should be on your local machine after pulling the repo. \
If they are not, please download them here and try again \n \
https://drive.google.com/uc?export=download&id=1fkrUbxsr3eGjjsfD3TcUOV8K0041N3FL\
\nOptionally, you may need to use pip install xlrd")

df_merged.head(10)

### 2. Data cleaning

# Check for missing values
df_merged.isnull().sum(axis=0)

### These missing values are caused by police districts not filing with the FBI. There is nothing we can do about this so we drop these observations

# Missing values caused by missing report
df_merged = df_merged[df_merged['TotalOffenses'].notna()]

Some population values are missing. We will fill these values with the population from another column.

# Fill in missing values
df_merged['Population1'] = df_merged['Population1'].fillna(df_merged['Population'])

df_merged = df_merged.iloc[:, :4]
df_merged.columns = ['City', 'Population', 'TotalOffenses', 'State']

The format of some state names are not consistent with the other datasets. We will need to fix this.

# Fix State names
df_merged['State'] = [' '.join(i.split('_')) for i in df_merged['State']]

# Check for missing values again
df_merged.isnull().sum(axis=0)

df_merged.head(20)

No missing values in the data. We will save it as a csv file.

### 3. Saving cleaned dataset

df_merged.to_csv('../data/data_cleaned/crime2020.csv', index=False)

### References

* Data source: https://cde.ucr.cjis.gov/LATEST/webapp/#/pages/home
* Read excel: https://pandas.pydata.org/docs/reference/api/pandas.read_excel.html
* https://stackoverflow.com/questions/20908018/import-multiple-excel-files-into-python-pandas-and-concatenate-them-into-one-dat
