==> 5_merge_crime_ACS.ipynb <==
# 90803 Data Cleaning and Question Definition
# Merge Crime Data (city level) with ACS data (city level)
#### The city level of the crime data and ACS data don't match up exactly so we will have to try our best with string matching on the city names

**Team 14**

Colton Lapp & Chi-Shiun Tsai

#import modules 
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import geopandas as gpd
from census import Census
from us import states
import numpy as np
import os
import seaborn as sns
import us
import re

# Read in Crime Data
This should be on your local machine from step 4

#Read in DFs
base_dir = '/'.join( os.getcwd().split('/')[0:-1] )
base_dir
crime = pd.read_csv(base_dir+ '/data/data_cleaned/crime2020.csv')
crime.head()

## Drop the word "township", "village", "city", etc from city names in crime data
These words are messing up our merges because they are inconsistent across datsets so we will drop them both

crime['City'] = crime['City'].apply(  lambda x: re.sub(" Township", "", x,flags=re.I) )
crime['City'] = crime['City'].apply(  lambda x: re.sub(" Town", "", x,flags=re.I) )
crime['City'] = crime['City'].apply(  lambda x: re.sub(" City", "", x,flags=re.I) )
crime['City'] = crime['City'].apply(  lambda x: re.sub(" Village", "", x,flags=re.I) )
crime['City'] = crime['City'].apply(  lambda x: re.sub("Metropolitan ", "", x,flags=re.I) )
crime['City'] = crime['City'].apply(  lambda x: re.sub(" Police Department", "", x,flags=re.I) )
crime['City'] = crime['City'].apply(  lambda x: re.sub("-Mecklenburg", "", x,flags=re.I) )
crime['City'] = crime['City'].apply(  lambda x: re.sub("-Clarke County", "", x,flags=re.I) )



crime['City'] = crime['City'].apply(  lambda x: x.strip() )
crime.head()

## Create "full name" column combining city and state

crime['full_name'] = crime['City'] + ', ' + crime['State']

# Read in Places shape file to merge onto crime data
#### This should be on your machine from step 1

places_shp = gpd.read_file( base_dir + '/data/shape_files/CensusPlaces')
places_shp.head(1)

### Remove "city", "Village", "town", etc from NAME variable in Census Shapefiles to make merges better

places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub(" Township", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub(" Town", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub(" City", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub(" Township", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub(" Village", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub(" CDP", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub("\(balance\)", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub("-Davidson metropolitan government", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub("-Fayette", "", x,flags=re.I) )
places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: re.sub("-Clarke County unified government", "", x,flags=re.I) )

places_shp['NAME'] = places_shp['NAME'].apply(  lambda x: x.strip() )

#Add state name to shape file
places_shp['State Name'] = places_shp['STATEFP'].apply( lambda x: str(us.states.lookup(x)) )
places_shp.head(1)

#Create "Full name" by combining state and city name
places_shp['Full_name_1'] =  places_shp['NAME'] + ', ' + places_shp['State Name']
places_shp['Full_name_2'] =  places_shp['NAMELSAD'] + ', ' + places_shp['State Name']
places_shp.head(3)

# Join crime to city shapefile by name (imperfect strategy)

merged = places_shp.merge(crime, left_on='Full_name_1', right_on='full_name', how='outer')


## Show Crime Observations that didn't get matched

pd.set_option('display.max_rows', 100)

missing = merged[ merged['STATEFP'].isnull() ].copy()

#See how many were not merged
print( "Total Crime observations that were not merged: ", missing.shape[0], "out of", crime.shape[0])
missing.sort_values('Population', ascending=False)[['City', 'State', 'Population']].head(20)

# Do Inner join of crime and city data
This is the best we can do with improperly matching cities. We did our best to manually look through the data and correct mistakes

#Do Inner Join and ignore missing matches
merged = places_shp.merge(crime, left_on='Full_name_1', right_on='full_name', how='inner')
merged = merged[['STATEFP', 'PLACEFP', 'GEOID', 'NAMELSAD', 'State Name', 'City', 'TotalOffenses']]

# Save Dataset

merged.to_csv( base_dir + '/data/data_cleaned/crime_cities.csv', index=False)

# References
1. https://www.statology.org/pandas-merge-on-different-column-names/
2. https://stackoverflow.com/questions/8813265/why-doesnt-ignorecase-flag-re-i-work-in-re-sub
3. https://stackoverflow.com/questions/38773379/simplest-python-equivalent-to-rs-gsub
4. https://www.geeksforgeeks.org/select-rows-that-contain-specific-text-using-pandas/#
