==> 6_merge_all_data.ipynb <==
### 90803 Data Cleaning and Question Definition
# Merge all data (ACS, crime, covid) together

**Team 14**

Colton Lapp & Chi-Shiun Tsai

#import modules 
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import geopandas as gpd
from census import Census
from us import states
import numpy as np
import os
import seaborn as sns
import us
import re

# 1. Read in all cleaned datsets

base_dir = os.getcwd().split('/')
base_dir = '/'.join(base_dir[0:-1])

ACS_data = pd.read_csv( base_dir + '/data/data_cleaned/ACS_5yr_data.csv')
covid_data = pd.read_csv( base_dir + '/data/data_cleaned/covid_cities.csv')
crime_data = pd.read_csv( base_dir + '/data/data_cleaned/crime_cities.csv')

print("Shape of city data: ", ACS_data.shape)
print("Shape of covid data: ", covid_data.shape)
print("Shape of crime data: ", crime_data.shape)

## 2. Merge all data together

#Merge all to ACS data
data_merged = ACS_data.merge(covid_data.drop('NAMELSAD', axis=1), on='GEOID', how='left')
data_merged = data_merged.merge(crime_data.drop(['STATEFP', 'PLACEFP', 'NAMELSAD', 'City'], axis=1), on='GEOID', how='left')

data_merged.info()
data_merged.head()

# 3. Subset correct predictor variables
There are some leftover variables from the shapefile merges we don't actually need, so drop them.

# Subset target variables
dta1 = data_merged.iloc[:, 0:43]
dta1.head(1)


# Subset target variables
dta2 = data_merged.iloc[:, 43:]
dta2 = dta2[['GEOID', 'Population 2019', 'Pop Chng %', 'ALAND', 'AWATER', 'cases_per_100K', 'deaths_per_100K']]
dta2.columns = ['GEOID', 'Population 2019', 'Pop Chng %', 'land_area', 'water_area', 'cases_per_100K', 'deaths_per_100K']
dta2.head(1)

dta = pd.concat([dta1, dta2], axis=1)
dta.head(5)

### Combine some ACS columns to make single variable
These variables need to be summed to be interpreted correctly because each individual variable only represents part of the total count.

# Combine columns
dta['Employed in Professional/Scientific/Technical Services'] = dta['Employed in Professional/Scientific/Technical Services'] + dta['Employed in Professional/Scientific/Technical Services 2']
dta['Employed in Accommodation/Food Services'] = dta['Employed in Accommodation/Food Services'] + dta['Employed in Accommodation/Food Services 2']
dta['Employed in Health Care/Social Assistance'] = dta['Employed in Health Care/Social Assistance'] + dta['Employed in Health Care/Social Assistance 2']
dta['Employed in Arts/Entertainment/Recreation'] = dta['Employed in Arts/Entertainment/Recreation'] + dta['Employed in Arts/Entertainment/Recreation 2']

dta.drop(['Employed in Professional/Scientific/Technical Services 2', 'Employed in Accommodation/Food Services 2',
           'Employed in Health Care/Social Assistance 2', 'Employed in Arts/Entertainment/Recreation 2'], axis=1, inplace=True)
dta.head(2)


# Check for missing values
dta.isnull().sum(axis=0)

# 4. Drop rows with missing observations
Some rows have missing values for COVID-19 data. Because the impact of COVID-19 is one of our main questions, we can't use these rows in our model, so we drop them.

data_complete = dta.dropna()
print("Number of observations: ", data_complete.shape)
data_complete.head(10)

# Check for missing values again to make sure that the data is complete
data_complete.isnull().sum(axis=0)

# 5. Save as csv

data_complete.to_csv('../data/data_cleaned/data_all.csv', index=False)

### References

1. https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.drop.html
2. https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.dropna.html
